requirements.txt
pytorch
cudatoolkit
recbole
openai
